+++ plugins/SebCore/src/SebCore/Properties/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("SebCore")]
 [assembly: AssemblyProduct("SebCore")]
-[assembly: AssemblyVersion("1.0.2")]
-[assembly: AssemblyFileVersion("1.0.2")]
+[assembly: AssemblyVersion("1.1.0")]
+[assembly: AssemblyFileVersion("1.1.0")]